backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # Required for file uploads & OAuth2 form data

# ── Database ───────────────────────────────────────────────────────
sqlalchemy==2.0.30
pymysql==1.1.1                 # MySQL driver used by SQLAlchemy
cryptography==42.0.7           # Required by PyMySQL for auth

# ── Authentication ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT encoding/decoding
bcrypt==4.1.3                      # Password hashing
passlib==1.7.4                     # Optional but safe to include

# ── AI / LLM ───────────────────────────────────────────────────────
openai==1.30.1                 # OpenAI client (used with Azure inference endpoint)
groq==0.9.0                    # Groq client for Whisper transcription

# ── Voice / Audio ──────────────────────────────────────────────────
# groq handles transcription — no extra audio library needed

# ── Date Utilities ─────────────────────────────────────────────────
python-dateutil==2.9.0         # relativedelta used in reminder recurrence logic

# ── Environment Variables ──────────────────────────────────────────
python-dotenv==1.0.1

# ── Pydantic (FastAPI uses this internally) ────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1